requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
psycopg2-binary==2.8.6
PyJWT==2.0.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
uuid==1.30
